+++ apps/data_sync_worker/cli.py
@@ -1,32 +1,14 @@
 import argparse
 import logging
 import os
-
+import deployment
 # from apps.data_sync_worker.main import run_synchronization # Example import
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 
 def main():
-    parser = argparse.ArgumentParser(description="Run the Data Synchronization Worker.")
-    parser.add_argument(
-        "--db-path",
-        type=str,
-        default="data/market_data.sqlite", # Default path for the worker to write to
-        help="Path to the market_data.sqlite database file to be created/updated.",
-    )
-    # Add other arguments if needed, e.g., specific tickers, date ranges for sync
-
-    args = parser.parse_args()
-
-    logger.info(f"Data Synchronization Worker started.")
-    logger.info(f"Database will be written to: {os.path.abspath(args.db_path)}")
-
-    
-    # Placeholder for actual sync logic
-    # run_synchronization(args.db_path) 
-    logger.info("Placeholder: Data synchronization logic would run here.")
-    logger.info(f"Data Synchronization Worker finished. DB should be at {os.path.abspath(args.db_path)}")
+    deployment.main()
 
 if __name__ == "__main__":
     main()
