--- apps/data_sync_worker/main.py
@@ -1,12 +1,0 @@
-import logging
-import sqlite3 # For database operations
-# Example: from packages.investor_agent_lib import yfinance_tools
-
-logger = logging.getLogger(__name__)
-# BasicConfig should ideally be in cli.py or a shared logging setup module
-# logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-
-
-def flow(db_path: str):
-    # Placeholder for database initialization logic
-    pass
