+++ config/paths.py
@@ -1,0 +1,8 @@
+from pathlib import Path
+
+# Project root directory (points to /Users/bytedance/work/aaw-financehub-mcp)
+PROJECT_ROOT = Path(__file__).parent
+
+# Data directories
+DATA_DIR = PROJECT_ROOT / "data"
+REPORT_PATH = DATA_DIR / "report"
