+++ data/tickers.txt
