+++ packages/investor_agent_lib/analytics/time_series.py
